+++ python/__init__.py
@@ -527,6 +527,8 @@ def trainingModel(model: Union['torch.nn.Module', 'poptorch.PoplarExecutor'],
     :returns: The :py:class:`poptorch.PoplarExecutor` wrapper to use in place
         of ``model``.
     """
+    local_options = copy.deepcopy(options) if options else Options()
+
     if isinstance(model, PoplarExecutor):
         model = model._user_model  # pylint: disable=protected-access
 
@@ -534,7 +536,7 @@ def trainingModel(model: Union['torch.nn.Module', 'poptorch.PoplarExecutor'],
     maybe_wrapped_model = copy.copy(model)
 
     return PoplarExecutor(model=maybe_wrapped_model,
-                          options=options,
+                          options=local_options,
                           training=True,
                           optimizer=optimizer,
                           user_model=model,
@@ -571,10 +573,13 @@ def inferenceModel(model: Union['torch.nn.Module', 'poptorch.PoplarExecutor'],
     #    out = model(image_input)
     #    poptorch_model.outputs([out])
 
+    local_options = copy.deepcopy(options) if options else Options()
+
     if isinstance(model, PoplarExecutor):
         model = model._user_model  # pylint: disable=protected-access
+
     return PoplarExecutor(model=copy.copy(model),
-                          options=options,
+                          options=local_options,
                           training=False,
                           poptorch_version=__version__)